Extract shared base class for other d8 hydrological algorithms
